Makefile: support building libkrb5 with musl on Linux
Linking libkrb5 requires that cgo passes -static while linking
intermediate artifacts. This was previously impossible due to a
bug in the Go toolchain (golang/go#16651), which has now been
resolved. Remove our workaround for this issue, and enable
building/linking libkrb5 when building with musl on Linux.

Release note: None
