--- pkg/ccl/gssapiccl/empty.go
@@ -1,12 +1,0 @@
-// Copyright 2018 The Cockroach Authors.
-//
-// Licensed as a CockroachDB Enterprise file under the Cockroach Community
-// License (the "License"); you may not use this file except in compliance with
-// the License. You may obtain a copy of the License at
-//
-//     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
-
-package gssapiccl
-
-// This file is here so on platforms where gssapi.go isn't built there's at
-// least one .go file to build.
